Refactor stubbed requests - subjects
Add spec helper for stubbing requests for subjects.
