+++ spec/features/result_filters/study_type_spec.rb
@@ -2,13 +2,14 @@ require 'rails_helper'
 
 describe 'Study type filter', type: :feature do
   include StubbedRequests::Courses
+  include StubbedRequests::Subjects
 
   let(:filter_page) { PageObjects::Page::ResultFilters::StudyType.new }
   let(:results_page) { PageObjects::Page::Results.new }
   let(:base_parameters) { results_page_parameters }
 
   before do
-    stub_subjects_request
+    stub_subjects
   end
 
   describe 'Study type filter page' do
