+++ spec/features/suggested_salary_searches_spec.rb
@@ -2,22 +2,13 @@ require 'rails_helper'
 
 describe 'Suggested salary searches', type: :feature do
   include StubbedRequests::Courses
+  include StubbedRequests::Subjects
 
   let(:filter_page) { PageObjects::Page::ResultFilters::Funding.new }
   let(:results_page) { PageObjects::Page::Results.new }
 
   def default_query_for_location_search(radius:)
     { l: 1, loc: 'Cat Town', lq: 'Cat Town', lat: 51.4980188, lng: -0.1300436, rad: radius }
-  end
-
-  def stub_subjects
-    stub_request(
-      :get,
-      "#{Settings.teacher_training_api.base_url}/api/v3/subjects?fields%5Bsubjects%5D=subject_name,subject_code&sort=subject_name",
-    ).to_return(
-      body: File.new('spec/fixtures/api_responses/subjects_sorted_name_code.json'),
-      headers: { "Content-Type": 'application/vnd.api+json; charset=utf-8' },
-    )
   end
 
   def suggested_search_count_parameters
