+++ spec/view_objects/results_view_spec.rb
@@ -2,6 +2,7 @@ require 'rails_helper'
 
 describe ResultsView do
   include StubbedRequests::Courses
+  include StubbedRequests::Subjects
 
   let(:query_parameters) { ActionController::Parameters.new(parameter_hash) }
 
@@ -16,13 +17,7 @@ describe ResultsView do
   end
 
   before do
-    stub_request(
-      :get,
-      "#{Settings.teacher_training_api.base_url}/api/v3/subjects?fields%5Bsubjects%5D=subject_name,subject_code&sort=subject_name",
-    ).to_return(
-      body: File.new('spec/fixtures/api_responses/subjects_sorted_name_code.json'),
-      headers: { "Content-Type": 'application/vnd.api+json; charset=utf-8' },
-    )
+    stub_subjects
   end
 
   describe '#query_parameters_with_defaults' do
